Code cleanup. v29 ready
